Certification fullpage layout (#4650)
Moves Certification to react #4608
