+++ app/views/certifications/v2.html.erb
@@ -1,3 +1,9 @@
+<% content_for :full_page_content do %>
 <%= react_component("Certification", props: {
-  vacolsId: @certification.vacols_id
+    userDisplayName: current_user.display_name,
+    dropdownUrls: dropdown_urls,
+    feedbackUrl: feedback_url,
+    buildDate: build_date,
+    vacolsId: @certification.vacols_id
 }) %>
+<% end %>
