+++ client/app/certification/components/ErrorMessage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import StatusMessage from '../components/StatusMessage';
+import StatusMessage from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/StatusMessage';
 
 const ErrorMessage = () => {
